server/pyproject.toml
[tool.poetry]
name = "noisekiller"
version = "0.1.0"
description = "High-performance AI Verification Layer"
authors = ["Shourjyo Chakraborty"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"